Create OtherMember.java
So I may have not realized that Matt was assigned the task of creating the Member and Family Classes and accidentally created my own versions of the Classes. I didn't feel like scrapping this work so I decided to just uploaded it here anyway. Sorry for accidentally stealing your task Matt! :)